refactor: remove static methods to use private methods
