+++ src/main/java/com/grego/linux_images_server/service/impl/ImageServiceImpl.java
@@ -4,20 +4,21 @@ import com.grego.linux_images_server.models.Image;
 import com.grego.linux_images_server.service.ImageService;
 import com.grego.linux_images_server.utils.Verifier;
 import lombok.extern.log4j.Log4j2;
+import net.coobird.thumbnailator.Thumbnails;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 
 import java.io.File;
 import java.io.IOException;
+import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Objects;
 import java.util.stream.Collectors;
 
-import static com.grego.linux_images_server.utils.FileNameManipulations.getFileExtension;
-import static com.grego.linux_images_server.utils.ThumbnailControl.generateThumbnail;
-import static com.grego.linux_images_server.utils.ThumbnailControl.getThumbnailName;
+
 
 @Log4j2
 @Service
@@ -54,12 +55,39 @@ public class ImageServiceImpl implements ImageService {
         String thumbSrc = IMAGE_DIRECTORY + getThumbnailName(fileName, THUMBNAIL_SUFFIX);
 
         try {
-            generateThumbnail(file.toPath(), uploadDirectory, THUMBNAIL_SUFFIX);
+            this.generateThumbnail(file.toPath(), uploadDirectory, THUMBNAIL_SUFFIX);
         } catch (IOException e) {
             log.error(e);
         }
 
         return new Image(fileType, id, fileName, src, thumbSrc);
+    }
+
+    private void generateThumbnail(Path imagePath, String uploadDirectory, String suffix) throws IOException {
+
+        String fileName = imagePath.getFileName().toString();
+        String thumbnailFileName = this.getThumbnailName(fileName, suffix);
+        Path thumbnailPath = Paths.get(uploadDirectory).resolve(thumbnailFileName);
+        Thumbnails.of(imagePath.toFile())
+                .size(200, 200)
+                .toFile(thumbnailPath.toFile());
+
+    }
+
+    private String getThumbnailName(String originalFileName, String suffix) {
+        int lastIndexOfDot = originalFileName.lastIndexOf('.');
+        if (lastIndexOfDot == -1) {
+            return originalFileName + suffix;
+        }
+        return originalFileName.substring(0, lastIndexOfDot) + suffix + originalFileName.substring(lastIndexOfDot);
+    }
+
+    private String getFileExtension(String fileName) {
+        int lastIndexOfDot = fileName.lastIndexOf('.');
+        if (lastIndexOfDot == -1) {
+            return "";
+        }
+        return fileName.substring(lastIndexOfDot + 1);
     }
